\name{POMaSPU-package}
\alias{POMaSPU-package}
\docType{package}
\title{
Adaptive Association Tests for Multiple Phenotypes using Proportional Odds Model (POM)
}
\description{
Provides adaptive association tests for multiple phenotypes.
}
\details{
POM-aSPU test evaluates an association between an ordinal response (categorical variable) and multiple variables (a set of phenotypes). 
We treat single SNP as an ordinal response and 
multiple traits as predictors in a proportional odds model (POM).
The proposed method can be applied to a high dimensional setting where the number of phenotypes (\emph{p}) can be larger than the sample size (\emph{n}), in addition to the usual small (\emph{p}) setting.}
\author{
Junghi Kim and Wei Pan
}
\references{
Kim J, and Pan W (2017). "Adaptive testing for multiple traits in a proportional odds model with applications to detect SNP-brain network associations." \emph{Genetic Epidemiology}, 41(3), 259--277.
}

